+++ app/src/config.rs
@@ -1,8 +1,8 @@
 pub struct Config {
     pub filename: String,
     pub base_dir: String,
-    pub prefix: String,
     pub vendor: String,
+    pub prefix: String,
 }
 
 impl Config {
@@ -13,12 +13,20 @@ impl Config {
 
         let filename = args[1].clone();
         let base_dir = args[2].clone();
+        let vendor = match args.get(3) {
+            Some(arg) => arg,
+            None => "App",
+        };
+        let prefix = match args.get(4) {
+            Some(arg) => arg,
+            None => "",
+        };
 
         Ok(Config {
             filename,
             base_dir,
-            prefix: String::from("Pre"),
-            vendor: String::from("App"),
+            vendor: vendor.to_string(),
+            prefix: prefix.to_string(),
         })
     }
 }